+++ src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/GuiceyExtensionsSupport.java
@@ -4,6 +4,7 @@ import com.google.common.base.Preconditions;
 import com.google.inject.Injector;
 import io.dropwizard.testing.DropwizardTestSupport;
 import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
@@ -21,8 +22,10 @@ import ru.vyarus.dropwizard.guice.test.util.HooksUtil;
 import ru.vyarus.dropwizard.guice.test.util.TestSetupUtils;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -31,11 +34,11 @@ import java.util.stream.Collectors;
  * Base class for junit 5 extensions implementations. All extensions use {@link DropwizardTestSupport} object
  * for actual execution (only configuration differs).
  * <p>
- * Extensions assumed to be used only on class level: extension will start dropwizard app before all tests
- * and shut down it after all tests. If nested tests used - they also affected. Execution per test is not allowed
- * because these tests are integration tests and they must minimize environment preparation time. Group tests
- * not affecting application state into one class and use different test classes (or nested classes) for tests
- * modifying state.
+ * Extensions might be used on class level (annotation and manual registration in static field; when extension start
+ * dropwizard app before all tests and shut down it after all tests) and on method level (manual registration in non
+ * static field; application starts before each test).
+ * <p>
+ * Nested tests also supported.
  * <p>
  * Test instance is not managed by guice! Only {@link com.google.inject.Injector#injectMembers(Object)} applied
  * for it to process test fields injection. Guice AOP can't be used on test methods. Technically, creating test
@@ -45,6 +48,9 @@ import java.util.stream.Collectors;
  * and initialize these hooks automatically. It was done like this to simplify customizations, when main extension
  * could be declared as annotation and hook as field. Also, it was impossible to implement hooks support
  * with junit extension. Hook field could be declared even in base test class.
+ * <p>
+ * Also, detects {@link ru.vyarus.dropwizard.guice.test.jupiter.env.TestEnvironmentSetup} fields annotated with
+ * {@link ru.vyarus.dropwizard.guice.test.jupiter.env.EnableSetup}. Behaviour is the same as with hook fields.
  * <p>
  * For external integrations (other extensions), there is a special "hack" allowing to access
  * {@link DropwizardTestSupport} object (and so get access to injector): {@link #lookupSupport(ExtensionContext)}.
@@ -57,7 +63,8 @@ import java.util.stream.Collectors;
 public abstract class GuiceyExtensionsSupport extends TestParametersSupport implements
         BeforeAllCallback,
         AfterAllCallback,
-        BeforeEachCallback {
+        BeforeEachCallback,
+        AfterEachCallback {
 
     // dropwizard support storage key (store visible for all relative tests)
     private static final String DW_SUPPORT = "DW_SUPPORT";
@@ -65,6 +72,8 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
     private static final String DW_CLIENT = "DW_CLIENT";
     // indicator storage key of nested test (when extension activated in parent test)
     private static final String INHERITED_DW_SUPPORT = "INHERITED_DW_SUPPORT";
+    // indicator storage key for case when application started for each method in test
+    private static final String PER_METHOD_DW_SUPPORT = "PER_METHOD_DW_SUPPORT";
 
     protected final TestExtensionsTracker tracker;
 
@@ -74,22 +83,8 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
 
     @Override
     public void beforeAll(final ExtensionContext context) throws Exception {
-        final ExtensionContext.Store store = getExtensionStore(context);
-        if (store.get(DW_SUPPORT) == null) {
-            // find fields annotated with @EnableHook and @EnableSetup
-            final FieldSupport fields = new FieldSupport(context.getRequiredTestClass(), tracker);
-            fields.activateBaseHooks();
-
-            final DropwizardTestSupport<?> support = prepareTestSupport(context, fields.getSetupObjects());
-            // activate hooks declared in test static fields (so hooks declared in annotation goes before)
-            fields.activateClassHooks();
-            store.put(DW_SUPPORT, support);
-            // for pure guicey tests client may seem redundant, but it can be used for calling other services
-            store.put(DW_CLIENT, new ClientSupport(support));
-
-            tracker.logExtensionRegistrations();
-
-            support.before();
+        if (!lookupSupport(context).isPresent()) {
+            start(context, null);
         } else {
             // in case of nested test, beforeAll for root extension will be called second time (because junit keeps
             // only one extension instance!) and this means we should not perform initialization, but we also must
@@ -105,7 +100,14 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
     }
 
     @Override
-    public void beforeEach(final ExtensionContext context) {
+    public void beforeEach(final ExtensionContext context) throws Exception {
+        // run-per-method support (activated with @RegisterExtension on non-static field only)
+        if (!lookupSupport(context).isPresent()) {
+            start(context, context.getTestInstance().get());
+            // mark per-method mode to properly shut down after test method
+            getLocalExtensionStore(context).put(PER_METHOD_DW_SUPPORT, true);
+        }
+
         // before each used to properly handle both default @TestInstance(TestInstance.Lifecycle.PER_METHOD)
         // and @TestInstance(TestInstance.Lifecycle.PER_CLASS) (in later case BeforeAllCallback called after
         // TestInstancePostProcessor, making it not usable for this task)
@@ -123,24 +125,25 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
     }
 
     @Override
+    public void afterEach(final ExtensionContext context) throws Exception {
+        if (getLocalExtensionStore(context).get(PER_METHOD_DW_SUPPORT) != null) {
+            stop(context);
+        }
+    }
+
+    @Override
     public void afterAll(final ExtensionContext context) throws Exception {
-        // just in case, normally hooks cleared automatically after appliance
-        ConfigurationHooksSupport.reset();
+        // do nothing in application per test method mode
+        if (lookupSupport(context).isPresent()) {
 
-        final ExtensionContext.Store localExtensionStore = getLocalExtensionStore(context);
-        if (localExtensionStore.get(INHERITED_DW_SUPPORT) != null) {
-            localExtensionStore.remove(INHERITED_DW_SUPPORT);
-            // do nothing: extension managed on upper context
-            return;
-        }
+            // nested tests support
+            final Object nestedTestMarker = getLocalExtensionStore(context).remove(INHERITED_DW_SUPPORT);
+            if (nestedTestMarker != null) {
+                // do nothing: extension managed on upper context
+                return;
+            }
 
-        final DropwizardTestSupport<?> support = getSupport(context);
-        if (support != null) {
-            support.after();
-        }
-        final ClientSupport client = getClient(context);
-        if (client != null) {
-            client.close();
+            stop(context);
         }
     }
 
@@ -191,7 +194,7 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
      * according to annotation.
      *
      * @param context extension context
-     * @param setups setup extensions resolved from fields (or empty list)
+     * @param setups  setup extensions resolved from fields (or empty list)
      * @return configured dropwizard test support object
      */
     protected abstract DropwizardTestSupport<?> prepareTestSupport(ExtensionContext context,
@@ -223,6 +226,38 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
                 .create(GuiceyExtensionsSupport.class, context.getRequiredTestClass()));
     }
 
+    private void start(final ExtensionContext context, final Object testInstance) throws Exception {
+        final ExtensionContext.Store store = getExtensionStore(context);
+        // find fields annotated with @EnableHook and @EnableSetup
+        final FieldSupport fields = new FieldSupport(context.getRequiredTestClass(), testInstance, tracker);
+        fields.activateBaseHooks();
+
+        final DropwizardTestSupport<?> support = prepareTestSupport(context, fields.getSetupObjects());
+        // activate hooks declared in test static fields (so hooks declared in annotation goes before)
+        fields.activateClassHooks();
+        store.put(DW_SUPPORT, support);
+        // for pure guicey tests client may seem redundant, but it can be used for calling other services
+        store.put(DW_CLIENT, new ClientSupport(support));
+
+        tracker.logExtensionRegistrations();
+
+        support.before();
+    }
+
+    private void stop(final ExtensionContext context) throws Exception {
+        // just in case, normally hooks cleared automatically after appliance
+        ConfigurationHooksSupport.reset();
+
+        final DropwizardTestSupport<?> support = getSupport(context);
+        if (support != null) {
+            support.after();
+        }
+        final ClientSupport client = getClient(context);
+        if (client != null) {
+            client.close();
+        }
+    }
+
     /**
      * Utility class for activating hooks and setup objects collected from fields (annotated with
      * {@link EnableHook} and {link {@link ru.vyarus.dropwizard.guice.test.jupiter.env.EnableSetup}}).
@@ -233,38 +268,42 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
      * Setup extensions from fields are always registered after all other.
      */
     private static class FieldSupport {
+        // test instance in application per test method case (beforeEach)
+        private final Object instance;
         private final TestExtensionsTracker tracker;
         private final List<Field> parentHookFields;
         private final List<Field> ownHookFields;
         private final List<Field> extensionFields;
 
-        FieldSupport(final Class<?> testClass, final TestExtensionsTracker tracker) {
+        FieldSupport(final Class<?> testClass, final Object instance, final TestExtensionsTracker tracker) {
+            this.instance = instance;
             this.tracker = tracker;
             // find and validate all fields
-            ownHookFields = findHookFields(testClass);
+            final boolean includeInstanceFields = instance != null;
+            ownHookFields = findHookFields(testClass, includeInstanceFields);
             parentHookFields = ownHookFields.isEmpty() ? Collections.emptyList() : ownHookFields.stream()
                     .filter(field -> !testClass.equals(field.getDeclaringClass()))
                     .collect(Collectors.toList());
             ownHookFields.removeAll(parentHookFields);
 
-            extensionFields = findSetupFields(testClass);
+            extensionFields = findSetupFields(testClass, includeInstanceFields);
         }
 
         public List<TestEnvironmentSetup> getSetupObjects() {
-            tracker.extensionsFromFields(extensionFields);
+            tracker.extensionsFromFields(extensionFields, instance);
             return extensionFields.isEmpty() ? Collections.emptyList() : getFieldValues(extensionFields);
         }
 
         public void activateBaseHooks() {
             // activate hooks declared in base classes
             activateFieldHooks(parentHookFields);
-            tracker.hooksFromFields(parentHookFields, true);
+            tracker.hooksFromFields(parentHookFields, true, instance);
         }
 
         public void activateClassHooks() {
             // activate all remaining hooks (in test class)
             activateFieldHooks(ownHookFields);
-            tracker.hooksFromFields(ownHookFields, false);
+            tracker.hooksFromFields(ownHookFields, false, instance);
         }
 
         private void activateFieldHooks(final List<Field> fields) {
@@ -274,18 +313,23 @@ public abstract class GuiceyExtensionsSupport extends TestParametersSupport impl
         @SuppressWarnings("unchecked")
         private <T> List<T> getFieldValues(final List<Field> fields) {
             return fields.isEmpty() ? Collections.emptyList() : (List<T>)
-                    ReflectionUtils.readFieldValues(fields, null);
+                    ReflectionUtils.readFieldValues(fields, instance);
         }
 
-        private List<Field> findHookFields(final Class<?> testClass) {
-            final List<Field> fields = AnnotationSupport.findAnnotatedFields(testClass, EnableHook.class);
-            HooksUtil.validateFieldHooks(fields);
+        private List<Field> findHookFields(final Class<?> testClass, final boolean includeInstanceFields) {
+            List<Field> fields = AnnotationSupport.findAnnotatedFields(testClass, EnableHook.class);
+            if (includeInstanceFields) {
+                fields = new ArrayList<>(fields); // original list is unmodifiable
+                // sort static fields first
+                fields.sort(Comparator.comparing(field -> Modifier.isStatic(field.getModifiers()) ? 0 : 1));
+            }
+            HooksUtil.validateFieldHooks(fields, includeInstanceFields);
             return fields.isEmpty() ? Collections.emptyList() : new ArrayList<>(fields);
         }
 
-        private List<Field> findSetupFields(final Class<?> testClass) {
+        private List<Field> findSetupFields(final Class<?> testClass, final boolean includeInstanceFields) {
             final List<Field> fields = AnnotationSupport.findAnnotatedFields(testClass, EnableSetup.class);
-            TestSetupUtils.validateFields(fields);
+            TestSetupUtils.validateFields(fields, includeInstanceFields);
             return fields.isEmpty() ? Collections.emptyList() : new ArrayList<>(fields);
         }
     }